refactor(125077): :hammer: move SolicitacoesEscola para outro arquivo
