+++ sme_sigpae_api/paineis_consolidados/solicitacoes_escola/api/viewsets.py
@@ -1,0 +1,836 @@
+import datetime
+import unicodedata
+
+from dateutil.relativedelta import relativedelta
+from django.db.models import Q
+from rest_framework import status
+from rest_framework.decorators import action
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+
+from sme_sigpae_api.cardapio.api.serializers.serializers import (
+    VinculoTipoAlimentoPeriodoSerializer,
+)
+from sme_sigpae_api.cardapio.models import (
+    VinculoTipoAlimentacaoComPeriodoEscolarETipoUnidadeEscolar,
+)
+from sme_sigpae_api.dados_comuns.permissions import PermissaoParaRecuperarDietaEspecial
+from sme_sigpae_api.escola.models import Escola, PeriodoEscolar
+from sme_sigpae_api.inclusao_alimentacao.models import GrupoInclusaoAlimentacaoNormal
+from sme_sigpae_api.kit_lanche.models import SolicitacaoKitLancheUnificada
+from sme_sigpae_api.medicao_inicial.models import SolicitacaoMedicaoInicial
+from sme_sigpae_api.paineis_consolidados.api.constants import (
+    AGUARDANDO_INICIO_VIGENCIA_DIETA_ESPECIAL,
+    ALTERACOES_ALIMENTACAO_AUTORIZADAS,
+    AUTORIZADAS_TEMPORARIAMENTE_DIETA_ESPECIAL,
+    AUTORIZADOS,
+    AUTORIZADOS_DIETA_ESPECIAL,
+    CANCELADOS,
+    CANCELADOS_DIETA_ESPECIAL,
+    CEU_GESTAO_PERIODOS_COM_SOLICITACOES_AUTORIZADAS,
+    FILTRO_ESCOLA_UUID,
+    INATIVAS_DIETA_ESPECIAL,
+    INATIVAS_TEMPORARIAMENTE_DIETA_ESPECIAL,
+    INCLUSOES_AUTORIZADAS,
+    INCLUSOES_ETEC_AUTORIZADAS,
+    KIT_LANCHES_AUTORIZADAS,
+    NEGADOS,
+    NEGADOS_DIETA_ESPECIAL,
+    PENDENTES_AUTORIZACAO,
+    PENDENTES_AUTORIZACAO_DIETA_ESPECIAL,
+    PESQUISA,
+    RELATORIO_PERIODO,
+    RELATORIO_RESUMO_MES_ANO,
+    RESUMO_ANO,
+    RESUMO_MES,
+    SUSPENSOES_AUTORIZADAS,
+)
+from sme_sigpae_api.paineis_consolidados.api.serializers import SolicitacoesSerializer
+from sme_sigpae_api.paineis_consolidados.api.viewsets import SolicitacoesViewSet
+from sme_sigpae_api.paineis_consolidados.models import SolicitacoesEscola
+from sme_sigpae_api.paineis_consolidados.utils.utils import (
+    formata_resultado_inclusoes_etec_autorizadas,
+    get_numero_alunos_alteracao_alimentacao,
+    tratar_append_return_dict,
+    tratar_data_evento_final_no_mes,
+    tratar_dias_duplicados,
+    tratar_inclusao_continua,
+    tratar_periodo_parcial,
+    tratar_periodo_parcial_cemei,
+)
+from sme_sigpae_api.paineis_consolidados.validators import FiltroValidator
+from sme_sigpae_api.relatorios.relatorios import (
+    relatorio_filtro_periodo,
+    relatorio_resumo_anual_e_mensal,
+)
+
+
+class EscolaSolicitacoesViewSet(SolicitacoesViewSet):
+    lookup_field = "uuid"
+    queryset = SolicitacoesEscola.objects.all()
+    permission_classes = (
+        IsAuthenticated,
+        PermissaoParaRecuperarDietaEspecial,
+    )
+    serializer_class = SolicitacoesSerializer
+
+    @action(detail=False, methods=["GET"], url_path=f"{PENDENTES_AUTORIZACAO}")
+    def pendentes_autorizacao(self, request):
+        escola_uuid = request.user.vinculo_atual.instituicao.uuid
+        query_set = SolicitacoesEscola.get_pendentes_autorizacao(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{PENDENTES_AUTORIZACAO_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def pendentes_autorizacao_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_pendentes_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{AUTORIZADOS_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def autorizados_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_autorizados_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{AUTORIZADAS_TEMPORARIAMENTE_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def autorizadas_temporariamente_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_autorizadas_temporariamente_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{AGUARDANDO_INICIO_VIGENCIA_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def aguardando_inicio_vigencia_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_aguardando_vigencia_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{INATIVAS_TEMPORARIAMENTE_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def inativas_temporariamente_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_inativas_temporariamente_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{INATIVAS_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+        permission_classes=(
+            IsAuthenticated,
+            PermissaoParaRecuperarDietaEspecial,
+        ),
+    )
+    def inativas_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_inativas_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{NEGADOS_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def negados_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_negados_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{CANCELADOS_DIETA_ESPECIAL}/{FILTRO_ESCOLA_UUID}",
+    )
+    def cancelados_dieta_especial(self, request, escola_uuid=None):
+        tem_parametro_sem_paginacao = request.GET.get("sem_paginacao", False)
+        query_set = SolicitacoesEscola.get_cancelados_dieta_especial(
+            escola_uuid=escola_uuid
+        )
+        query_set = SolicitacoesEscola.busca_filtro_dietas_especiais(
+            query_set, request.query_params
+        )
+        if tem_parametro_sem_paginacao:
+            return self._retorno_base(query_set, True)
+        return self._retorno_base(query_set)
+
+    @action(detail=False, methods=["GET"], url_path=f"{AUTORIZADOS}")
+    def autorizados(self, request):
+        escola_uuid = request.user.vinculo_atual.instituicao.uuid
+        query_set = SolicitacoesEscola.get_autorizados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        return self._retorno_base(query_set)
+
+    @action(
+        detail=False,
+        methods=("get",),
+        url_path=f"{CEU_GESTAO_PERIODOS_COM_SOLICITACOES_AUTORIZADAS}",
+    )
+    def ceu_gestao_periodos_com_solicitacoes_autorizadas(self, request):
+        escola_uuid = request.query_params.get("escola_uuid")
+        mes = request.query_params.get("mes")
+        ano = request.query_params.get("ano")
+
+        uuids_inclusoes_normais = GrupoInclusaoAlimentacaoNormal.objects.filter(
+            status="CODAE_AUTORIZADO",
+            escola__uuid=escola_uuid,
+            inclusoes_normais__cancelado=False,
+            inclusoes_normais__data__month=mes,
+            inclusoes_normais__data__year=ano,
+            inclusoes_normais__data__lt=datetime.date.today(),
+        ).values_list("uuid", flat=True)
+
+        periodos_escolares_inclusoes = PeriodoEscolar.objects.filter(
+            quantidadeporperiodo__grupo_inclusao_normal__uuid__in=uuids_inclusoes_normais
+        ).distinct()
+        escola = Escola.objects.get(uuid=escola_uuid)
+        vinculos = (
+            VinculoTipoAlimentacaoComPeriodoEscolarETipoUnidadeEscolar.objects.filter(
+                periodo_escolar__in=periodos_escolares_inclusoes,
+                ativo=True,
+                tipo_unidade_escolar=escola.tipo_unidade,
+            ).order_by("periodo_escolar__posicao")
+        )
+
+        return Response(
+            VinculoTipoAlimentoPeriodoSerializer(vinculos, many=True).data,
+            status=status.HTTP_200_OK,
+        )
+
+    @action(detail=False, methods=["GET"], url_path=f"{INCLUSOES_AUTORIZADAS}")
+    def inclusoes_autorizadas(self, request):  # noqa C901
+        escola_uuid = request.query_params.get("escola_uuid")
+        mes = request.query_params.get("mes")
+        ano = request.query_params.get("ano")
+        date = datetime.date(int(ano), int(mes), 1)
+        periodos_escolares = request.query_params.getlist("periodos_escolares[]")
+
+        query_set = SolicitacoesEscola.get_autorizados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        query_set = query_set.filter(
+            Q(data_evento__month=mes, data_evento__year=ano)
+            | Q(data_evento__lt=date, data_evento_2__gte=date)
+        )
+        query_set = query_set.filter(data_evento__lt=datetime.date.today())
+        query_set = self.remove_duplicados_do_query_set(query_set)
+        sol_medicao_inicial = SolicitacaoMedicaoInicial.objects.filter(
+            escola__uuid=escola_uuid, mes=mes, ano=ano
+        ).first()
+        return_dict = []
+        for inclusao in query_set:
+            inc = inclusao.get_raw_model.objects.get(uuid=inclusao.uuid)
+            if inclusao.tipo_doc == "INC_ALIMENTA_CEI":
+                periodos_internos = []
+                periodos_externos = []
+                if "PARCIAL" in periodos_escolares:
+                    periodos_externos = ["INTEGRAL"]
+                    periodos_internos = ["MANHA", "TARDE"]
+                if "INTEGRAL" in periodos_escolares:
+                    periodos_externos = ["INTEGRAL"]
+                    periodos_internos = ["INTEGRAL"]
+                if "MANHA" in periodos_escolares:
+                    periodos_externos = ["MANHA"]
+                    periodos_internos = ["MANHA"]
+                if "TARDE" in periodos_escolares:
+                    periodos_externos = ["TARDE"]
+                    periodos_internos = ["TARDE"]
+                dias_motivos = inc.dias_motivos_da_inclusao_cei.filter(
+                    data__month=mes, data__year=ano, cancelado=False
+                )
+                quantidade_por_faixa = inc.quantidade_alunos_da_inclusao.filter(
+                    periodo__nome__in=periodos_internos,
+                    periodo_externo__nome__in=periodos_externos,
+                )
+                if quantidade_por_faixa:
+                    for dia_motivo in dias_motivos:
+                        faixas_etarias_uuids = quantidade_por_faixa.values_list(
+                            "faixa_etaria__uuid", flat=True
+                        )
+                        return_dict.append(
+                            {
+                                "dia": dia_motivo.data.day,
+                                "faixas_etarias": faixas_etarias_uuids.distinct(),
+                            }
+                        )
+            elif inclusao.tipo_doc == "INC_ALIMENTA_CEMEI":
+                dias_motivos_cemei = inc.dias_motivos_da_inclusao_cemei.filter(
+                    data__month=mes, data__year=ano
+                )
+                for periodo in periodos_escolares:
+                    if "Infantil" not in periodo:
+                        eh_parcial_integral = None
+                        if not inc.quantidade_alunos_cei_da_inclusao_cemei.exists():
+                            continue
+                        if periodo == "PARCIAL":
+                            qtd_alunos_cei_cemei_por_periodo = (
+                                inc.quantidade_alunos_cei_da_inclusao_cemei.filter(
+                                    periodo_escolar__nome__in=["MANHA", "TARDE"]
+                                )
+                            )
+                            eh_parcial_integral = False
+                            if not qtd_alunos_cei_cemei_por_periodo.exists():
+                                qtd_alunos_cei_cemei_por_periodo = (
+                                    inc.quantidade_alunos_cei_da_inclusao_cemei.filter(
+                                        periodo_escolar__nome="INTEGRAL"
+                                    )
+                                )
+                                eh_parcial_integral = True
+                        else:
+                            if (
+                                periodo == "INTEGRAL"
+                                and sol_medicao_inicial
+                                and not sol_medicao_inicial.ue_possui_alunos_periodo_parcial
+                            ):
+                                qtd_alunos_cei_cemei_por_periodo = (
+                                    inc.quantidade_alunos_cei_da_inclusao_cemei.filter(
+                                        periodo_escolar__nome__in=[
+                                            "INTEGRAL",
+                                            "MANHA",
+                                            "TARDE",
+                                        ]
+                                    )
+                                )
+                            else:
+                                qtd_alunos_cei_cemei_por_periodo = (
+                                    inc.quantidade_alunos_cei_da_inclusao_cemei.filter(
+                                        periodo_escolar__nome=periodo
+                                    )
+                                )
+                        if not qtd_alunos_cei_cemei_por_periodo.exists():
+                            continue
+                        faixas_etarias_uuids = (
+                            qtd_alunos_cei_cemei_por_periodo.values_list(
+                                "faixa_etaria__uuid", flat=True
+                            )
+                        )
+                        for dia_motivo_cemei in dias_motivos_cemei:
+                            if [
+                                r_dict
+                                for r_dict in return_dict
+                                if r_dict["dia"] == dia_motivo_cemei.data.day
+                            ]:
+                                if [
+                                    r_dict
+                                    for r_dict in return_dict
+                                    if r_dict["dia"] == dia_motivo_cemei.data.day
+                                    and r_dict["eh_parcial_integral"]
+                                ] and not eh_parcial_integral:
+                                    return_dict = [
+                                        r_dict
+                                        for r_dict in return_dict
+                                        if r_dict["dia"] != dia_motivo_cemei.data.day
+                                    ]
+                                    return_dict.append(
+                                        {
+                                            "dia": dia_motivo_cemei.data.day,
+                                            "faixas_etarias": faixas_etarias_uuids.distinct(),
+                                            "eh_parcial_integral": eh_parcial_integral,
+                                        }
+                                    )
+                            else:
+                                return_dict.append(
+                                    {
+                                        "dia": dia_motivo_cemei.data.day,
+                                        "faixas_etarias": faixas_etarias_uuids.distinct(),
+                                        "eh_parcial_integral": eh_parcial_integral,
+                                    }
+                                )
+                    else:
+                        periodo = periodo.split(" ")[1]
+                        if not inc.quantidade_alunos_emei_da_inclusao_cemei.filter(
+                            periodo_escolar__nome=periodo
+                        ).exists():
+                            continue
+                        for dia_motivo_cemei in dias_motivos_cemei:
+                            tratar_append_return_dict(
+                                dia_motivo_cemei.data.day,
+                                mes,
+                                ano,
+                                inc.quantidade_alunos_emei_da_inclusao_cemei.get(
+                                    periodo_escolar__nome=periodo
+                                ),
+                                inclusao,
+                                return_dict,
+                            )
+            else:
+                for periodo in inc.quantidades_periodo.all():
+                    if periodo.periodo_escolar.nome in periodos_escolares:
+                        if inclusao.tipo_doc == "INC_ALIMENTA_CONTINUA":
+                            if not periodo.cancelado:
+                                tratar_inclusao_continua(
+                                    mes, ano, periodo, inclusao, return_dict
+                                )
+                        else:
+                            for inclusao_normal in inc.inclusoes_normais.filter(
+                                data__month=mes, data__year=ano, cancelado=False
+                            ):
+                                tratar_append_return_dict(
+                                    inclusao_normal.data.day,
+                                    mes,
+                                    ano,
+                                    periodo,
+                                    inclusao,
+                                    return_dict,
+                                )
+        data = {"results": return_dict}
+
+        return Response(data)
+
+    @action(detail=False, methods=["GET"], url_path=f"{SUSPENSOES_AUTORIZADAS}")
+    def suspensoes_autorizadas(self, request):
+        escola_uuid = request.query_params.get("escola_uuid")
+        mes = request.query_params.get("mes")
+        ano = request.query_params.get("ano")
+        nome_periodo_escolar = request.query_params.get("nome_periodo_escolar")
+
+        query_set = SolicitacoesEscola.get_autorizados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        query_set = query_set.filter(data_evento__month=mes, data_evento__year=ano)
+        query_set = query_set.filter(data_evento__lt=datetime.date.today())
+        query_set = self.remove_duplicados_do_query_set(query_set)
+        return_dict = []
+
+        for suspensao in query_set:
+            susp = suspensao.get_raw_model.objects.get(uuid=suspensao.uuid)
+            if susp.DESCRICAO == "Suspensão de Alimentação de CEI":
+                nome_periodo_escolar = tratar_periodo_parcial(nome_periodo_escolar)
+                if nome_periodo_escolar in susp.periodos_escolares.all().values_list(
+                    "nome", flat=True
+                ):
+                    return_dict.append(
+                        {
+                            "dia": f"{susp.data.day:02d}",
+                            "periodo": nome_periodo_escolar,
+                            "motivo": susp.motivo.nome,
+                            "inclusao_id_externo": susp.id_externo,
+                        }
+                    )
+            else:
+                nome_periodo_escolar = tratar_periodo_parcial_cemei(
+                    nome_periodo_escolar, susp
+                )
+                s_quant_por_periodo = susp.quantidades_por_periodo.filter(
+                    periodo_escolar__nome=nome_periodo_escolar
+                )
+                for s_quant_periodo in s_quant_por_periodo:
+                    for suspensao in susp.suspensoes_alimentacao.all():
+                        tipos_alimentacao = s_quant_periodo.tipos_alimentacao.all()
+                        alimentacoes = [
+                            unicodedata.normalize(
+                                "NFD", alimentacao.nome.replace(" ", "_")
+                            )
+                            .encode("ascii", "ignore")
+                            .decode("utf-8")
+                            .lower()
+                            for alimentacao in tipos_alimentacao
+                        ]
+                        return_dict.append(
+                            {
+                                "dia": f"{suspensao.data.day:02d}",
+                                "periodo": nome_periodo_escolar,
+                                "alimentacoes": alimentacoes,
+                                "numero_alunos": s_quant_periodo.numero_alunos,
+                                "inclusao_id_externo": susp.id_externo,
+                            }
+                        )
+
+        data = {"results": return_dict}
+
+        return Response(data)
+
+    def trata_lanche_emergencial_queryset(self, eh_lanche_emergencial, query_set):
+        if eh_lanche_emergencial == "true":
+            query_set = query_set.filter(motivo__icontains="Emergencial")
+        else:
+            query_set = query_set.exclude(motivo__icontains="Emergencial")
+        return query_set
+
+    def get_alteracao_obj(self, alteracao, nome_periodo_escolar):
+        alt = None
+        if alteracao.escola.eh_cemei:
+            if "Infantil" not in nome_periodo_escolar:
+                return alt
+            nome_periodo_escolar = nome_periodo_escolar.split(" ")[1]
+            if alteracao.substituicoes_cemei_emei_periodo_escolar.filter(
+                periodo_escolar__nome=nome_periodo_escolar
+            ).exists():
+                alt = alteracao.substituicoes_cemei_emei_periodo_escolar.get(
+                    periodo_escolar__nome=nome_periodo_escolar
+                )
+        elif alteracao.substituicoes_periodo_escolar.filter(
+            periodo_escolar__nome=nome_periodo_escolar
+        ).exists():
+            alt = alteracao.substituicoes_periodo_escolar.get(
+                periodo_escolar__nome=nome_periodo_escolar
+            )
+        return alt
+
+    def alteracoes_lanche_emergencial(
+        self,
+        eh_lanche_emergencial,
+        alteracao,
+        alteracao_alimentacao,
+        mes,
+        ano,
+        return_dict,
+    ):
+        if eh_lanche_emergencial == "true":
+            for data_evento in alteracao.datas_intervalo.filter(
+                data__month=mes, data__year=ano, cancelado=False
+            ):
+                return_dict.append(
+                    {
+                        "dia": f"{data_evento.data.day:02d}",
+                        "numero_alunos": get_numero_alunos_alteracao_alimentacao(
+                            alteracao
+                        ),
+                        "inclusao_id_externo": alteracao.id_externo,
+                        "motivo": alteracao_alimentacao.motivo,
+                    }
+                )
+        return return_dict
+
+    def alteracoes_RPL_LPR(
+        self,
+        eh_lanche_emergencial,
+        alteracao,
+        alteracao_alimentacao,
+        nome_periodo_escolar,
+        mes,
+        ano,
+        return_dict,
+    ):
+        if eh_lanche_emergencial != "true":
+            alt = self.get_alteracao_obj(alteracao, nome_periodo_escolar)
+            if alt:
+                for data_evento in alteracao.datas_intervalo.filter(
+                    data__month=mes, data__year=ano, cancelado=False
+                ):
+                    return_dict.append(
+                        {
+                            "dia": f"{data_evento.data.day:02d}",
+                            "periodo": nome_periodo_escolar,
+                            "numero_alunos": alt.qtd_alunos,
+                            "inclusao_id_externo": alteracao.id_externo,
+                            "motivo": alteracao_alimentacao.motivo,
+                        }
+                    )
+        return return_dict
+
+    @action(
+        detail=False, methods=["GET"], url_path=f"{ALTERACOES_ALIMENTACAO_AUTORIZADAS}"
+    )
+    def alteracoes_alimentacoes_autorizadas(self, request):
+        escola_uuid = request.query_params.get("escola_uuid")
+        escola = Escola.objects.get(uuid=escola_uuid)
+
+        if escola.eh_cei:
+            return Response({"results": []}, status=status.HTTP_200_OK)
+
+        mes = request.query_params.get("mes")
+        ano = request.query_params.get("ano")
+        nome_periodo_escolar = request.query_params.get("nome_periodo_escolar")
+        eh_lanche_emergencial = request.query_params.get("eh_lanche_emergencial", "")
+
+        query_set = SolicitacoesEscola.get_autorizados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        query_set = query_set.filter(data_evento__month=mes, data_evento__year=ano)
+        query_set = query_set.filter(data_evento__lt=datetime.date.today())
+        query_set = self.trata_lanche_emergencial_queryset(
+            eh_lanche_emergencial, query_set
+        )
+        query_set = self.remove_duplicados_do_query_set(query_set)
+        return_dict = []
+
+        for alteracao_alimentacao in query_set:
+            alteracao = alteracao_alimentacao.get_raw_model.objects.get(
+                uuid=alteracao_alimentacao.uuid
+            )
+            return_dict = self.alteracoes_lanche_emergencial(
+                eh_lanche_emergencial,
+                alteracao,
+                alteracao_alimentacao,
+                mes,
+                ano,
+                return_dict,
+            )
+            return_dict = self.alteracoes_RPL_LPR(
+                eh_lanche_emergencial,
+                alteracao,
+                alteracao_alimentacao,
+                nome_periodo_escolar,
+                mes,
+                ano,
+                return_dict,
+            )
+
+        data = {"results": return_dict}
+
+        return Response(data)
+
+    @action(detail=False, methods=["GET"], url_path=f"{KIT_LANCHES_AUTORIZADAS}")
+    def kit_lanches_autorizadas(self, request):
+        escola_uuid = request.query_params.get("escola_uuid")
+        mes = request.query_params.get("mes")
+        ano = request.query_params.get("ano")
+
+        query_set = SolicitacoesEscola.get_autorizados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        query_set = query_set.filter(data_evento__month=mes, data_evento__year=ano)
+        query_set = query_set.filter(data_evento__lt=datetime.date.today())
+        query_set = self.remove_duplicados_do_query_set(query_set)
+        return_dict = []
+
+        for kit_lanche in query_set:
+            kit_lanche = kit_lanche.get_raw_model.objects.get(uuid=kit_lanche.uuid)
+            if kit_lanche:
+                if kit_lanche.DESCRICAO == "Kit Lanche CEMEI":
+                    dia = f"{kit_lanche.data.day:02d}"
+                    numero_alunos = kit_lanche.total_kits
+                else:
+                    dia = f"{kit_lanche.solicitacao_kit_lanche.data.day:02d}"
+                    numero_alunos = (
+                        kit_lanche.total_kit_lanche_escola(escola_uuid)
+                        if isinstance(kit_lanche, SolicitacaoKitLancheUnificada)
+                        else kit_lanche.quantidade_alimentacoes
+                    )
+                return_dict.append(
+                    {
+                        "dia": dia,
+                        "numero_alunos": numero_alunos,
+                        "kit_lanche_id_externo": kit_lanche.id_externo,
+                    }
+                )
+
+        data = {"results": return_dict}
+
+        return Response(data)
+
+    @action(detail=False, methods=["GET"], url_path=f"{INCLUSOES_ETEC_AUTORIZADAS}")
+    def inclusoes_etec_autorizadas(self, request):
+        escola_uuid = request.query_params.get("escola_uuid")
+        mes = request.query_params.get("mes")
+        ano = request.query_params.get("ano")
+        date = datetime.date(int(ano), int(mes), 1)
+
+        query_set = SolicitacoesEscola.get_autorizados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        query_set = query_set.filter(
+            Q(data_evento__month=mes, data_evento__year=ano)
+            | Q(data_evento__lt=date, data_evento_2__gte=date)
+        )
+        query_set = query_set.filter(
+            data_evento__lt=datetime.date.today(), motivo="ETEC"
+        )
+        query_set = self.remove_duplicados_do_query_set(query_set)
+
+        return_dict = []
+
+        def append(dia, inclusao):
+            resultado = formata_resultado_inclusoes_etec_autorizadas(
+                dia, mes, ano, inclusao
+            )
+            return_dict.append(resultado) if resultado else None
+
+        for sol_escola in query_set:
+            inclusao = sol_escola.get_raw_model.objects.get(uuid=sol_escola.uuid)
+            dia = sol_escola.data_evento.day
+            big_range = False
+            data_evento_final_no_mes = None
+            if sol_escola.data_evento.month != int(
+                mes
+            ) and sol_escola.data_evento_2.month != int(mes):
+                big_range = True
+                i = datetime.date(int(ano), int(mes), 1)
+                data_evento_final_no_mes = (i + relativedelta(day=31)).day
+                dia = datetime.date(int(ano), int(mes), 1).day
+            elif sol_escola.data_evento.month != int(mes):
+                big_range = True
+                data_evento_final_no_mes = sol_escola.data_evento_2.day
+                dia = datetime.date(int(ano), int(mes), 1).day
+            else:
+                data_evento_final_no_mes = sol_escola.data_evento_2.day
+            data_evento_final_no_mes = tratar_data_evento_final_no_mes(
+                data_evento_final_no_mes, sol_escola, big_range
+            )
+            while dia <= data_evento_final_no_mes:
+                append(dia, inclusao)
+                dia += 1
+        data = {"results": tratar_dias_duplicados(return_dict)}
+
+        return Response(data)
+
+    @action(detail=False, methods=["GET"], url_path=f"{NEGADOS}")
+    def negados(self, request):
+        escola_uuid = request.user.vinculo_atual.instituicao.uuid
+        query_set = SolicitacoesEscola.get_negados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        return self._retorno_base(query_set)
+
+    @action(detail=False, methods=["GET"], url_path=f"{CANCELADOS}")
+    def cancelados(self, request):
+        escola_uuid = request.user.vinculo_atual.instituicao.uuid
+        query_set = SolicitacoesEscola.get_cancelados(escola_uuid=escola_uuid)
+        query_set = SolicitacoesEscola.busca_filtro(query_set, request.query_params)
+        return self._retorno_base(query_set)
+
+    @action(detail=False, methods=["GET"], url_path=f"{RESUMO_ANO}")
+    def evolucao_solicitacoes(self, request):
+        usuario = request.user
+        escola_uuid = usuario.vinculo_atual.instituicao.uuid
+        query_set = SolicitacoesEscola.get_solicitacoes_ano_corrente(
+            escola_uuid=escola_uuid
+        )
+        response = {
+            "results": self._agrupa_por_mes_por_solicitacao(query_set=query_set)
+        }
+        return Response(response)
+
+    @action(detail=False, methods=["GET"], url_path=f"{RESUMO_MES}")
+    def resumo_mes(self, request):
+        usuario = request.user
+        escola_uuid = usuario.vinculo_atual.instituicao.uuid
+        totais_dict = SolicitacoesEscola.resumo_totais_mes(
+            escola_uuid=escola_uuid,
+        )
+        return Response(totais_dict)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{RELATORIO_RESUMO_MES_ANO}",
+    )
+    def relatorio_resumo_anual_e_mensal(self, request):
+        usuario = request.user
+        escola_uuid = usuario.vinculo_atual.instituicao.uuid
+
+        query_set = SolicitacoesEscola.get_solicitacoes_ano_corrente(
+            escola_uuid=escola_uuid
+        )
+        resumo_do_ano = self._agrupa_por_mes_por_solicitacao(query_set=query_set)
+        resumo_do_mes = SolicitacoesEscola.resumo_totais_mes(
+            escola_uuid=escola_uuid,
+        )
+        return relatorio_resumo_anual_e_mensal(request, resumo_do_mes, resumo_do_ano)
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        url_path=f"{RELATORIO_PERIODO}",
+    )
+    def relatorio_filtro_periodo(self, request):
+        usuario = request.user
+        escola = usuario.vinculo_atual.instituicao
+        form = FiltroValidator(request.GET)
+        if form.is_valid():
+            cleaned_data = form.cleaned_data
+            query_set = SolicitacoesEscola.filtros_escola(
+                escola_uuid=escola.uuid,
+                data_inicial=cleaned_data.get("data_inicial"),
+                data_final=cleaned_data.get("data_final"),
+                tipo_solicitacao=cleaned_data.get("tipo_solicitacao"),
+                status_solicitacao=cleaned_data.get("status_solicitacao"),
+            )
+            query_set = self.remove_duplicados_do_query_set(query_set)
+
+            return relatorio_filtro_periodo(
+                request, query_set, escola.nome, escola.diretoria_regional.nome
+            )
+        else:
+            return Response(form.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    @action(detail=False, methods=["GET"], url_path=f"{PESQUISA}")
+    def filtro_periodo_tipo_solicitacao(self, request):
+        """Filtro de todas as solicitações da escola.
+
+        ---
+        tipo_solicitacao -- ALT_CARDAPIO|INV_CARDAPIO|INC_ALIMENTA|INC_ALIMENTA_CONTINUA|
+        KIT_LANCHE_AVULSA|SUSP_ALIMENTACAO|KIT_LANCHE_UNIFICADA|TODOS
+        status_solicitacao -- AUTORIZADOS|NEGADOS|CANCELADOS|RECEBIDAS|TODOS
+        data_inicial -- dd-mm-yyyy
+        data_final -- dd-mm-yyyy
+        """
+        usuario = request.user
+        escola_uuid = usuario.vinculo_atual.instituicao.uuid
+        form = FiltroValidator(request.GET)
+        if form.is_valid():
+            cleaned_data = form.cleaned_data
+            query_set = SolicitacoesEscola.filtros_escola(
+                escola_uuid=escola_uuid,
+                data_inicial=cleaned_data.get("data_inicial"),
+                data_final=cleaned_data.get("data_final"),
+                tipo_solicitacao=cleaned_data.get("tipo_solicitacao"),
+                status_solicitacao=cleaned_data.get("status_solicitacao"),
+            )
+            return self._retorno_base(query_set)
+        else:
+            return Response(form.errors, status=status.HTTP_400_BAD_REQUEST)
